Split camera into two separate classes. Use "position" class to update and determine player position. "Camera" class is for rendering only.
